6 (#1)
* [bugfix][apperance] Patch dummy.html (#52)

* Update issue templates

* Prettified Code!

* [feature] Major Update which includes Dark Mode Toggle. (#46)


- Revert the background on the Navbar and Footer as it can cause people to have sensory issues

- Remove the button from the footer on all pages (not just index.html) (#42)

- Add a space between the Navbar and the Ukraine Banner #44 

- Fix colors of some links

- Add Dark Mode Toggle

Authored-by: Mbrick2 <Mbrick2@users.noreply.github.com>

* Rename bug_report.md to bug_report_template.md

* Create blank_issue.md

* Update issue templates

* [bugfix] Fix issues adressed in #59

* Created CodeQL

Added this since it would point out errors. If something wrong happens, then we can revert it easily.

* Update index.html

* Changed a class

Found a class that wasn't in the CSS code so I removed a word from it.

* Added CSS

Hopefully, I was able to change the "Learn More" on the Ukraine notice. If not, then I will have a mental breakdown, as all programmers do when they are stuck in coding.

* Update index.html

* Changed the "AspectOS" in the copyright sign since the text was white in light mode

* Changed some classes with the 'AspectOS' on the footer (copyright)

* Changed some code for the copyright date

* Reverted a commit where the code for the copyright date was changed

* Changed some stuff to make dark mode work on the copyright date. Now, time to work on the entire footer

* [github] New Bug Report System

* Update main.css

* Changed some dark mode errors such as some text being black in dark mode and white in light mode

* Inserted Issue Templates

Added issue templates for organization

* Prettified Code!

* Deleted old logo

Deleted the old logo since we don't need it anymore. Will be used for future Aspect Inc. Website

* Deleted banner

Why do we need this? Besides, it has the old logo

* Added new logo

Added the new logo to the website since it is outdated at the moment

* Delete logo

Problem with the logo at the moment

* Added PNG version of the logo

It would actually load

* Update index.html

* Add files via upload

* Github Actions Test (#62)

Co-authored-by: Mbrick2 <90691415+Mbrick2@users.noreply.github.com>

* Changed Favicon for all pages

* Added @media for one class as a test for (#61)

* [bugfix] Fix Dark Mode Issues (#63)

* [bugfix] Fix some minor bugs & [meta] update content for new logo

* [bugfix] Daily Dose of Bug Fixes

* Added Favicon

Added Favicon for home page

* Changed typo

* Fixed some errors

Fixed random characters that were on the page

* Fixed some errors

Fixed random characters that were on the page

* Fixed some errors

Fixed random characters that were on the page

* Fixed some errors

Fixed random characters that were on the page

* Fixed some errors

Fixed random characters that were on the page

* Fixed some errors

Fixed random characters that were on the page

* Fixed some errors

Fixed random characters that were on the page

* Update index.html

* Update index.html

* Update index.html

* Updated Favicon

Didn't work the first time...

* Updated home page

Changed description of what we do

* Update index.html

updated the staff list

* updated staff list

* Removed some stuff

Removed since those links could cause a person's computer to freeze

* Delete bug_report.md

* Fixed Grammer

* Update index.html

* Capitalized 'The Future Starts Here' in the footer

Co-authored-by: yooit_32 <ajonathanalexis@hotmail.com>
Co-authored-by: AIGamesDeveloper <AIGamesDeveloper@users.noreply.github.com>
Co-authored-by: Mbrick2 <90691415+Mbrick2@users.noreply.github.com>
Co-authored-by: Dream-123-123 <97002382+DreamyCodes@users.noreply.github.com>